Refactoring of ResourceFinder. Moved common methods from ResourceFinder to Fixture. Moved ResourceToCopy to top-level class.
