+++ src/main/java/org/concordion/internal/ConcordionBuilder.java
@@ -13,7 +13,6 @@ import org.concordion.api.extension.ConcordionExtender;
 import org.concordion.api.extension.ConcordionExtension;
 import org.concordion.api.extension.ConcordionExtensionFactory;
 import org.concordion.api.listener.*;
-import org.concordion.internal.ResourceFinder.ResourceToCopy;
 import org.concordion.internal.command.*;
 import org.concordion.internal.listener.*;
 import org.concordion.internal.util.Announcer;
